JS divided into components
