+++ src/index.html
@@ -168,8 +168,7 @@
     </script>
 
   <script src="https://cdnjs.cloudflare.com/ajax/libs/handlebars.js/4.1.0/handlebars.min.js"></script>
-  <script src="js/functions.js"></script>
-  <script src="js/script.js"></script>
+  <script src="js/app.js" type="module"></script>
 </body>
 
 </html>
