+++ src/js/components/Cart.js
@@ -1,0 +1,152 @@
+import { settings, select, classNames, templates } from '../settings.js';
+import utils from '../utils.js';
+import CartProduct from './CartProduct.js';
+
+class Cart {
+  constructor(element) {
+    const thisCart = this;
+
+    thisCart.products = [];
+
+    thisCart.deliveryFee = settings.cart.defaultDeliveryFee;
+
+    thisCart.getElements(element);
+    thisCart.initActions();
+
+    // console.log('new Cart', thisCart);
+  }
+
+  getElements(element) {
+    const thisCart = this;
+
+    thisCart.dom = {};
+
+    thisCart.dom.wrapper = element;
+    thisCart.dom.toggleTrigger = thisCart.dom.wrapper.querySelector(select.cart.toggleTrigger);
+    thisCart.dom.productList = thisCart.dom.wrapper.querySelector(select.cart.productList);
+
+    thisCart.renderTotalsKeys = ['totalNumber', 'totalPrice', 'subtotalPrice', 'deliveryFee'];
+
+    for (let key of thisCart.renderTotalsKeys) {
+      thisCart.dom[key] = thisCart.dom.wrapper.querySelectorAll(select.cart[key]);
+    }
+
+    thisCart.dom.form = thisCart.dom.wrapper.querySelector(select.cart.form);
+    thisCart.dom.phone = thisCart.dom.wrapper.querySelector(select.cart.phone);
+    thisCart.dom.address = thisCart.dom.wrapper.querySelector(select.cart.address);
+  }
+
+  initActions() {
+    const thisCart = this;
+
+    thisCart.dom.toggleTrigger.addEventListener('click', function () {
+      thisCart.dom.wrapper.classList.toggle(classNames.cart.wrapperActive);
+    });
+
+    thisCart.dom.productList.addEventListener('updated', function () {
+      thisCart.update();
+    });
+
+    thisCart.dom.productList.addEventListener('remove', function () {
+      thisCart.remove(event.detail.cartProduct);
+    });
+
+    thisCart.dom.form.addEventListener('submit', function () {
+      event.preventDefault();
+      thisCart.sendOrder();
+    });
+
+  }
+
+  sendOrder() {
+    const thisCart = this;
+
+    const url = settings.db.url + '/' + settings.db.order;
+
+    const payload = {
+      address: thisCart.dom.address.value,
+      phone: thisCart.dom.phone.value,
+      totalNumber: thisCart.totalNumber,
+      subtotalPrice: thisCart.subtotalPrice,
+      deliveryFee: thisCart.deliveryFee,
+      totalPrice: thisCart.totalPrice,
+
+      products: [],
+    };
+
+    for (thisCart.product of thisCart.products) {
+      payload.products.push(thisCart.product.getData());
+    }
+
+    const options = {
+      method: 'POST',
+      headers: {
+        'Content-Type': 'application/json',
+      },
+      body: JSON.stringify(payload),
+    };
+
+    fetch(url, options)
+      .then(function (response) {
+        return response.json();
+      })
+      .then(function (parsedResponse) {
+        console.log('Zamówienie dodane!', parsedResponse);
+      });
+  }
+
+  update() {
+    const thisCart = this;
+
+    thisCart.totalNumber = 0;
+    thisCart.subtotalPrice = 0;
+
+    for (thisCart.product of thisCart.products) {
+      thisCart.subtotalPrice = thisCart.subtotalPrice + thisCart.product.price;
+      thisCart.totalNumber = thisCart.totalNumber + thisCart.product.amount;
+    }
+
+    thisCart.totalPrice = thisCart.subtotalPrice + thisCart.deliveryFee;
+
+    // console.log(thisCart.totalNumber, thisCart.subtotalPrice, this.totalPrice);
+
+    for (let key of thisCart.renderTotalsKeys) {
+      for (let elem of thisCart.dom[key]) {
+        elem.innerHTML = thisCart[key];
+      }
+    }
+  }
+
+  remove(cartProduct) {
+    const thisCart = this;
+
+    const index = thisCart.products.indexOf(cartProduct);
+    thisCart.products.splice(index, 1);
+    cartProduct.dom.wrapper.remove();
+
+    thisCart.update();
+  }
+
+  add(menuProduct) {
+    const thisCart = this;
+
+    /* generate HTML based on template */
+    const generatedHTML = templates.cartProduct(menuProduct);
+
+    /* create element using utils.createElementFromHTML */
+    const generatedDOM = utils.createDOMFromHTML(generatedHTML);
+
+    /* find cart conatiner */
+    const cartContainer = thisCart.dom.productList;
+
+    /*add element to cart */
+    cartContainer.appendChild(generatedDOM);
+
+    thisCart.products.push(new CartProduct(menuProduct, generatedDOM));
+    // console.log('thisCart.products', thisCart.products);
+
+    thisCart.update();
+  }
+}
+
+export default Cart;
