+++ src/js/utils.js
@@ -57,3 +57,5 @@ Handlebars.registerHelper('ifEquals', function (arg1, arg2, options) {
 Handlebars.registerHelper('joinValues', function (input, options) {
   return Object.values(input).join(options.fn(this));
 });
+
+export default utils;
